refactored caching system, need to get dataVisuals now working again
