+++ main.py
@@ -3,8 +3,9 @@ from website import create_app
 from website.views import viewRoutes
 from website.auth import authRoutes
 from website.dataVisuals import dataVisualsRoutes
-from website.db import DB
-from website.cacheRedis import cacheRedis
+from website.cachingService.cachingSystem import CachingSystem 
+from website.cachingService.cachingClient import CacheClient
+from website.cachingService.databaseClient import DatabaseClient
 
 # get mongoURI from txt file
 with open("mongoURI.txt", "r") as file:
@@ -16,20 +17,15 @@ class flaskAppWrapper:
         # create and setup the flask app
         self.__app = create_app()  
         self.__app.config["MONGO_URI"] = (mongoURI)
-
-        self.__cacheRedis = cacheRedis()
-
-        # Setup the mongoDB connection 
-        self.__mongoDB = DB(self.__app, mongoURI, cacheRedis())
-        self.__mongoDB.configureConnection()
+        self.__cachingSystem = CachingSystem(CacheClient(), DatabaseClient(self.__app, mongoURI))
 
         self.setupRoutes()  
         
     def setupRoutes(self):
         # create the different routes, and call their protected methods which are overridden from baseView abstract super class
         views = viewRoutes(self.__app)
-        auth = authRoutes(self.__app, self.__mongoDB, self.__cacheRedis)
-        dataVisuals = dataVisualsRoutes(self.__app, self.__mongoDB, self.__cacheRedis)
+        auth = authRoutes(self.__app, self.__cachingSystem)
+        dataVisuals = dataVisualsRoutes(self.__app, self.__cachingSystem)
         views._setupRoutes()
         auth._setupRoutes()
         dataVisuals._setupRoutes()
