+++ website/dataVisuals.py
@@ -1,8 +1,6 @@
 import sys
 import os
 
-from website.cacheRedis import cacheRedis
-
 # Add the project root directory to sys.path
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), "..")))
 
@@ -20,10 +18,9 @@ import matplotlib.pyplot as plt
 
 # wrap auth routes in a class for OOP
 class dataVisualsRoutes(baseView):
-    def __init__(self, flaskApp, mongoDB, cachRedis):
+    def __init__(self, flaskApp, cachingSystem):
         super().__init__(flaskApp)
-        self.__mongoDB = mongoDB
-        self.__cacheRedis = cacheRedis
+        self.__cachingSystem = cachingSystem
 
     def _setupRoutes(self):
         # setup routes
